Refactoring back from textbook-mvc
