--- src/main/java/doob/controller/LevelController.java
@@ -1,17 +1,0 @@
-package doob.controller;
-
-import java.util.ArrayList;
-
-import doob.model.Ball;
-import doob.model.Projectile;
-
-public class LevelController {
-	
-	private ArrayList<Ball> balls;
-	private ArrayList<Projectile> projectiles;
-	private int ballSpeed = 3;
-	private int shootSpeed = 12;
-	private int startHeight = 200;
-	private int ballSize = 100;
-	
-}
