+++ src/main/java/doob/controller/MenuController.java
@@ -3,6 +3,9 @@ package doob.controller;
 import doob.App;
 import javafx.fxml.FXML;
 
+/**
+ * Opening menu.
+ */
 public class MenuController {
 	
 	@FXML
