--- src/main/java/doob/model/PlayerModel.java
@@ -1,43 +1,0 @@
-package doob.model;
-
-import doob.controller.Collidable;
-import javafx.scene.canvas.GraphicsContext;
-import javafx.scene.image.Image;
-
-/**
- * PlayerModel class, handles.
- */
-public class PlayerModel {
-
-    public PlayerModel() {
-    }
-
-    /*public void move() {
-        this.x = x + speed;
-    }*/
-
-    /*
-    *//**
-     * Moves the character to the right.
-     *//*
-    public void moveRight() {
-        this.x = x + speed;
-        this.direction = Direction.RIGHT;
-    }
-
-    *//**
-     * Moves the character to the left.
-     *//*
-    public void moveLeft() {
-        this.x = x - speed;
-        this.direction = Direction.LEFT;
-    }
-
-    *//**
-     * The character stands in idle.
-     *//*
-    public void stand() {
-        this.direction = Direction.STAND;
-    }*/
-
-}
