+++ src/main/java/doob/model/Spike.java
@@ -2,7 +2,7 @@ package doob.model;
 
 import javafx.scene.image.Image;
 
-public class Spike extends Projectile{
+public class Spike extends Projectile {
 	
 	public Spike(double x, double y, double shootSpeed) {
 		super(x, y, shootSpeed);
