+++ src/main/java/doob/model/game/SinglePlayerGame.java
@@ -1,15 +1,15 @@
 package doob.model.game;
 
 import doob.model.Level;
-import doob.model.PlayerModel;
+import doob.model.Player;
 
 import java.util.ArrayList;
 
 public class SinglePlayerGame extends AbstractGame {
-	private PlayerModel player;
+	private Player player;
 
 	public SinglePlayerGame(int score, int difficulty, 
-			ArrayList<Level> levels, PlayerModel p1) {
+			ArrayList<Level> levels, Player p1) {
 		super(score, difficulty, levels);
 		this.player = p1;
 	}
